🎨 code cleanup and reorder
