+++ apps/contracts/src/tests/vault/one_strategy.ts
@@ -1,80 +1,21 @@
 import { Keypair } from "@stellar/stellar-sdk";
 import { AddressBook } from "../../utils/address_book.js";
-import { checkUserBalance } from "../strategy.js";
 import {
-  admin,
   CreateVaultParams,
-  deployVault,
   depositToVault,
-  fetchCurrentInvestedFunds,
-  fetchParsedCurrentIdleFunds,
   Instruction,
   manager,
   pauseStrategy,
-  queueVaultManager,
   rebalanceVault,
   rescueFromStrategy,
-  setEmergencyManager,
-  setFeeReceiver,
-  setRebalanceManager,
-  setVaultManager,
   unpauseStrategy,
-  upgradeVaultWasm,
   withdrawFromVault
 } from "../vault.js";
 import { green, purple, red, yellow } from "../common.js";
-import { airdropAccount, installContract } from "../../utils/contract.js";
-
-async function fetchBalances(addressBook: AddressBook, vault_address: string, user: Keypair) {
-  console.log(yellow, "---------------------------------------");
-  console.log(yellow, "Fetching balances");
-  console.log(yellow, "---------------------------------------");
-
-  const idle_funds = await fetchParsedCurrentIdleFunds(
-    vault_address,
-    user
-  );
-  const invested_funds = await fetchCurrentInvestedFunds(
-    vault_address,
-    user
-  );
-  const hodl_balance = await checkUserBalance(
-    addressBook.getContractId("hodl_strategy"),
-    vault_address,
-    user
-  );
-
-  return {idle_funds, invested_funds, hodl_balance};
-}
-async function deployOneStrategyVault(addressBook: AddressBook, params: CreateVaultParams[], user: Keypair) {
-  console.log(purple, "---------------------------------------");
-  console.log(purple, "Deploying vault with one strategy");
-  console.log(purple, "---------------------------------------");
-  try {
-    const { 
-      address: vault_address, 
-      instructions:deploy_instructions, 
-      readBytes:deploy_read_bytes, 
-      writeBytes:deploy_write_bytes
-    } = await deployVault(
-      addressBook,
-      params,
-      "TestVault",
-      "TSTV"
-    );
-    console.log(vault_address);
-    return {address: vault_address, deploy_instructions, deploy_read_bytes, deploy_write_bytes};
-  } catch (error:any) {
-    console.error(red, error);
-    return {
-      address: null, 
-      deploy_instructions:0, 
-      deploy_read_bytes: 0, 
-      deploy_write_bytes: 0,
-      error
-    };
-  }
-}
+import { airdropAccount } from "../../utils/contract.js";
+import { deployDefindexVault, fetchBalances } from "./utils.js";
+import { testAccessControl } from "./access_control.js";
+import { testUpgradeContract } from "./upgrade_contract.js";
 
 /* 
 // One asset one strategy success flow:
@@ -122,7 +63,7 @@ export async function oneAssetOneStrategySuccess(addressBook: AddressBook, param
     deploy_read_bytes,
     deploy_write_bytes 
 
-  } = await deployOneStrategyVault(addressBook, params, user);
+  } = await deployDefindexVault(addressBook, params);
   if (!vault_address) throw new Error("Vault was not deployed");
 
   const { 
@@ -1065,374 +1006,11 @@ export async function oneAssetOneStrategySuccess(addressBook: AddressBook, param
   return {tableData, budgetData};
 }
 
-/* 
-// Access control tests:
-  - [x] try setManager without previous queued manager
-  - [x] try queueManager from unauthorized
-  - [x] queueManager
-  - [x] try setManager before time is up
-  - [ ] try setManager from unauthorized
-  - [ ] setManager
-
-  - [x] try setRebalanceManager from unauthorized
-  - [x] setRebalanceManager
-
-  - [x] try setFeeReceiver from unauthorized
-  - [x] setFeeReceiver
-
-  - [x] try setEmergencyManager from unauthorized
-  - [x] setEmergencyManager
-*/
-async function testAccessControl(addressBook: AddressBook, params: CreateVaultParams[], user: Keypair) {
-  //Deploy vault
-  const { 
-    address:vault_address, 
-  } = await deployOneStrategyVault(addressBook, params, user);
-  if (!vault_address) throw new Error("Vault was not deployed");
-
-  //Try set manager without previous queued manager
-  try {
-    console.log(purple, "---------------------------------------");
-    console.log(purple, "Set manager");
-    console.log(purple, "---------------------------------------");
-    const { instructions, readBytes, writeBytes } = await setVaultManager(vault_address, user);
-    return { instructions, readBytes, writeBytes };
-  } catch (error: any ) {
-    if (error.toString().includes("HostError: Error(Contract, #134)")) {
-      console.log(green, "------------------------------------------------------------------");
-      console.log(green, "| Set manager without previous queued manager failed as expected |");
-      console.log(green, "------------------------------------------------------------------");
-    } else {
-      throw Error(error);
-    }
-  }
-
-  //Try queue manager from unauthorized
-  try {
-    console.log(purple, "---------------------------------------");
-    console.log(purple, "Try queue manager from unauthorized");
-    console.log(purple, "---------------------------------------");
-    const random_user = Keypair.random();
-    await airdropAccount(random_user);
-    const  {result} = await queueVaultManager(vault_address, random_user, random_user);
-    if( result !== false){
-      throw Error("Queue manager from unauthorized validation failed");
-    } else if (result === false) {
-      console.log(green, "------------------------------------------------------");
-      console.log(green, "| Queue manager from unauthorized failed as expected |");
-      console.log(green, "------------------------------------------------------");
-    }
-  } catch (error: any) {
-    throw Error(error);
-  }
-
-  const new_manager = Keypair.random();
-  await airdropAccount(new_manager);
-  // Queue manager
-  const {
-    result,
-    instructions: queue_instructions,
-    readBytes: queue_read_bytes,
-    writeBytes: queue_write_bytes
-  } = await (async ()=>{
-      try {
-        console.log(purple, "---------------------------------------");
-        console.log(purple, "Queue manager");
-        console.log(purple, "---------------------------------------");
-        const  {result, instructions, readBytes, writeBytes } = await queueVaultManager(vault_address, manager, new_manager);
-        if( result !== true){
-          console.log(green, "-----------------------------------");
-          console.log(green, "| Queue manager set successufully |");
-          console.log(green, "-----------------------------------");
-        } else if (result === false) {
-          console.error(red, "Queue manager from unauthorized validation failed");
-        }
-        return {result, instructions, readBytes, writeBytes};
-      } catch (error: any) {
-        console.error(red, error);
-        return {result: false, instructions: 0, readBytes: 0, writeBytes: 0};
-      }
-    }
-  )();
-
-  //Try setManager before time is up
-  try {
-    console.log(purple, "---------------------------------------");
-    console.log(purple, "Set manager before time");
-    console.log(purple, "---------------------------------------");
-    const { instructions, readBytes, writeBytes } = await setVaultManager(vault_address, user);
-    return { instructions, readBytes, writeBytes };
-  } catch (error: any ) {
-    if (error.toString().includes("HostError: Error(Contract, #133)")) {
-      console.log(green, "----------------------------------------------");
-      console.log(green, "| Set manager before time failed as expected |");
-      console.log(green, "----------------------------------------------");
-    } else {
-      throw Error(error);
-    }
-  } 
-
-  // Try setRebalanceManager from unauthorized
-  try {
-    console.log(purple, "---------------------------------------");
-    console.log(purple, "Try setRebalanceManager from unauthorized");
-    console.log(purple, "---------------------------------------");
-    const random_user = Keypair.random();
-    await airdropAccount(random_user);
-    const {result}  = await setRebalanceManager(vault_address, random_user, user.publicKey());
-    if( result !== false){
-      throw Error("Set rebalance manager from unauthorized validation failed");
-    } else if (result === false) {
-      console.log(green, "--------------------------------------------------------------");
-      console.log(green, "| Set rebalance manager from unauthorized failed as expected |");
-      console.log(green, "--------------------------------------------------------------");
-    }
-  } catch (error: any) {
-    throw Error(error);
-  }
-
-  // setRebalanceManager success
-  const {
-    instructions: set_rebalance_manager_instructions, 
-    readBytes: set_rebalance_manager_read_bytes, 
-    writeBytes: set_rebalance_manager_write_bytes
-  } = await (async () => {
-    try {
-      console.log(purple, "---------------------------------------");
-      console.log(purple, "setRebalanceManager");
-      console.log(purple, "---------------------------------------");
-      const random_user = Keypair.random();
-      await airdropAccount(random_user);
-      const {result, instructions, readBytes, writeBytes}  = await setRebalanceManager(vault_address, manager, random_user.publicKey());
-      if( result === null){
-        console.log(green, "--------------------------------------");
-        console.log(green, "| Rebalance manager set sucessfully  |");
-        console.log(green, "--------------------------------------");
-      } else if (result === false) {
-        throw Error("Set rebalance manager failed");
-      }
-      return {result, instructions, readBytes, writeBytes};
-    } catch (error: any) {
-      console.error(red, error);
-      return {result: false, instructions: 0, readBytes: 0, writeBytes: 0};
-    } 
-  } )();
-
-  // Try set fee reciever from unauthorized
-  try {
-    console.log(purple, "---------------------------------------");
-    console.log(purple, "Try set fee receiver from unauthorized");
-    console.log(purple, "---------------------------------------");
-    const random_user = Keypair.random();
-    await airdropAccount(random_user);
-    await setFeeReceiver(vault_address, random_user, user.publicKey());
-
-  } catch (error: any) {
-    if( error.toString().includes("HostError: Error(Contract, #130)")) {
-      console.log(green, "----------------------------------------------------------");
-      console.log(green, "| Set fee receiver from unauthorized failed as expected |");
-      console.log(green, "----------------------------------------------------------");
-    } else {     
-      throw Error(error);
-    }
-  } 
-
-  // setFeeReceiver success
-  const {
-    instructions: set_fee_receiver_instructions,
-    readBytes: set_fee_receiver_read_bytes,
-    writeBytes: set_fee_receiver_write_bytes
-  } = await (async () => {
-    try {
-      console.log(purple, "---------------------------------------");
-      console.log(purple, "setFeeReceiver");
-      console.log(purple, "---------------------------------------");
-      const {instructions, readBytes, writeBytes} = await setFeeReceiver(vault_address, manager, user.publicKey());
-      console.log(green, "---------------------------------");
-      console.log(green, "| Fee receiver set sucessfully  |");
-      console.log(green, "---------------------------------");
-      return {instructions, readBytes, writeBytes};
-    } catch (error: any) {
-      console.error(red, error);
-      return {instructions: 0, readBytes: 0, writeBytes: 0};
-    } 
-  } )();
- 
-
-  // Try set emergency manager from unauthorized
-  try {
-    console.log(purple, "-------------------------------------------");
-    console.log(purple, "Try set emergency manager from unauthorized");
-    console.log(purple, "-------------------------------------------");
-    const random_user = Keypair.random();
-    await airdropAccount(random_user);
-    const {result} = await setEmergencyManager(vault_address, random_user, user.publicKey());
-    if( result !== false){
-      throw Error("Set emergency manager from unauthorized validation failed");
-    } else if (result === false) {
-      console.log(green, "--------------------------------------------------------------");
-      console.log(green, "| Set emergency manager from unauthorized failed as expected |");
-      console.log(green, "--------------------------------------------------------------");
-    }
-  } catch (error: any) {
-    throw Error(error);
-  }
-
-  // setEmergencyManager success
-  const {
-    instructions: set_emergency_manager_instructions,
-    readBytes: set_emergency_manager_read_bytes,
-    writeBytes: set_emergency_manager_write_bytes
-  } = await (async () => {
-    try {
-      console.log(purple, "---------------------------------------");
-      console.log(purple, "setEmergencyManager");
-      console.log(purple, "---------------------------------------");
-      const random_user = Keypair.random();
-      await airdropAccount(random_user);
-      const {result, instructions, readBytes, writeBytes} = await setEmergencyManager(vault_address, manager, random_user.publicKey());
-      if( result === null){
-        console.log(green, "--------------------------------------");
-        console.log(green, "| Emergency manager set sucessfully  |");
-        console.log(green, "--------------------------------------");
-      } else if (result === false) {
-        throw Error("Set emergency manager failed");
-      }
-      return {result, instructions, readBytes, writeBytes};
-    } catch (error: any) {
-      console.error(red, error);
-      return {result: false, instructions: 0, readBytes: 0, writeBytes: 0};
-    }
-  } )();
-
-  const tests_status = {
-    queue: {
-      status: !!queue_instructions && !!queue_read_bytes && !!queue_write_bytes,
-    },
-    set_rebalance_manager: {
-      status: !!set_rebalance_manager_instructions && !!set_rebalance_manager_read_bytes && !!set_rebalance_manager_write_bytes,
-    },
-    set_fee_receiver: {
-      status: !!set_fee_receiver_instructions && !!set_fee_receiver_read_bytes && !!set_fee_receiver_write_bytes,
-    },
-    set_emergency_manager: {
-      status: !!set_emergency_manager_instructions && !!set_emergency_manager_read_bytes && !!set_emergency_manager_write_bytes,
-    }
-  }
-
-  const budgetData = {
-    queue: {
-      status: tests_status.queue.status ? `success`: `failed`,
-      instructions: queue_instructions,
-      readBytes: queue_read_bytes,
-      writeBytes: queue_write_bytes,
-    },
-    set_rebalance_manager: {
-      status: tests_status.set_rebalance_manager.status ? `success`: `failed`,
-      instructions: set_rebalance_manager_instructions,
-      readBytes: set_rebalance_manager_read_bytes,
-      writeBytes: set_rebalance_manager_write_bytes,
-    },
-    set_fee_receiver: {
-      status: tests_status.set_fee_receiver.status ? `success`: `failed`,
-      instructions: set_fee_receiver_instructions,
-      readBytes: set_fee_receiver_read_bytes,
-      writeBytes: set_fee_receiver_write_bytes,
-    },
-    set_emergency_manager: {
-      status: tests_status.set_emergency_manager.status ? `success`: `failed`,
-      instructions: set_emergency_manager_instructions,
-      readBytes: set_emergency_manager_read_bytes,
-      writeBytes: set_emergency_manager_write_bytes,
-    }
-  }
-
-  return {
-    budgetData
-  };
-};
-
-/* 
-// Upgrade tests:
-  - [x] try upgrade from unauthorized
-  - [x] upgrade
-*/
-export async function testUpgradeContract(addressBook: AddressBook, params: CreateVaultParams[], user: Keypair) {
-  try {
-    console.log(yellow, "----------------");
-    console.log(yellow, "| updating Wasm |");
-    console.log(yellow, "----------------");
-    await installContract("defindex_vault", addressBook, admin);
-    console.log(green, "----------------");
-    console.log(green, "| Wasm updated |");
-    console.log(green, "----------------");
-  } catch (error: any) {
-    throw Error(error);
-  }
-  
-  // Deploy vault
-  const { 
-    address:vault_address, 
-  } = await deployOneStrategyVault(addressBook, params, user);
-
-  if (!vault_address) throw new Error("Vault was not deployed");
-  //Try upgrade from unauthorized
-  try {
-    console.log(purple, "---------------------------------------");
-    console.log(purple, "Try upgrade from unauthorized");
-    console.log(purple, "---------------------------------------");
-    const random_user = Keypair.random();
-    const wasm_hash = Buffer.from(addressBook.getWasmHash("defindex_vault"), "hex");
-    await airdropAccount(random_user);
-    const {result} = await upgradeVaultWasm(vault_address, random_user, wasm_hash);
-    if( result !== false){
-      throw Error("Upgrade from unauthorized validation failed");
-    } else if (result === false) {
-      console.log(green, "------------------------------------------------");
-      console.log(green, "| Upgrade from unauthorized failed as expected |");
-      console.log(green, "------------------------------------------------");
-    }
-  } catch (error: any) {
-    throw Error(error);
-  }
-
-  // upgrade success
-  const {
-    instructions: upgrade_instructions,
-    readBytes: upgrade_read_bytes,
-    writeBytes: upgrade_write_bytes
-  } = await (async () => {
-    try {
-      console.log(purple, "---------------------------------------");
-      console.log(purple, "Upgrade");
-      console.log(purple, "---------------------------------------");
-      const wasm_hash = Buffer.from(addressBook.getWasmHash("defindex_vault"), "hex");
-      const {instructions, readBytes, writeBytes} = await upgradeVaultWasm(vault_address, manager, wasm_hash);
-      console.log(green, "------------------------");
-      console.log(green, "| Upgrade sucessfully  |");
-      console.log(green, "------------------------");
-      return {instructions, readBytes, writeBytes};
-    } catch (error: any) {
-      console.error(red, error);
-      return {instructions: 0, readBytes: 0, writeBytes: 0};
-    } 
-  } )();
-  const budgetData = {
-    upgrade: {
-      status: !!upgrade_instructions && !!upgrade_read_bytes && !!upgrade_write_bytes ? `success`: `failed`,
-      instructions: upgrade_instructions,
-      readBytes: upgrade_read_bytes,
-      writeBytes: upgrade_write_bytes,
-    }
-  }
-  return { budgetData };
-
-}
 
 export async function testVaultOneAssetOneStrategy(addressBook: AddressBook, params: CreateVaultParams[], user: Keypair) {
   const {tableData: userFlowTable, budgetData: userFlowBudgetData} = await oneAssetOneStrategySuccess(addressBook, params, user);
   const {budgetData: accessControlBudgetData} = await testAccessControl(addressBook, params, user);
-  const {budgetData: upgradeBudgetData} = await testUpgradeContract(addressBook, params, user);
+  const {budgetData: upgradeBudgetData} = await testUpgradeContract(addressBook, params);
 
   const tableData:any  = {...userFlowTable,};
   const budgetData:any = { ...userFlowBudgetData, ...accessControlBudgetData,  ...upgradeBudgetData};
